apps/assets/serializers.py
from rest_framework import serializers
from .models import Asset, ScanResult

class AssetSerializer(serializers.ModelSerializer):
    class Meta:
        model = Asset
        fields = "__all__"

class ScanResultSerializer(serializers.ModelSerializer):
    class Meta:
        model = ScanResult
        fields = "__all__"