requirements.txt
Django
djangorestframework
psycopg2-binary
python-dotenv
requests
apscheduler
supabase
psutil
screeninfo
pypiwin32